ContextReorganizationAction.runnable = True, moved to file with other agent-actions.
