+++ openhands/events/action/__init__.py
@@ -5,11 +5,11 @@ from openhands.events.action.agent import (
     AgentRejectAction,
     AgentThinkAction,
     ChangeAgentStateAction,
+    ContextReorganizationAction,
     RecallAction,
 )
 from openhands.events.action.browse import BrowseInteractiveAction, BrowseURLAction
 from openhands.events.action.commands import CmdRunAction, IPythonRunCellAction
-from openhands.events.action.context_reorganization import ContextReorganizationAction
 from openhands.events.action.empty import NullAction
 from openhands.events.action.files import (
     FileEditAction,
