+++ openhands/events/action/agent.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import Any
+from typing import Any, ClassVar, List
 
 from openhands.core.schema import ActionType
 from openhands.events.action.action import Action
@@ -195,3 +195,33 @@ class CondensationAction(Action):
         if self.summary:
             return f'Summary: {self.summary}'
         return f'Condenser is dropping the events: {self.forgotten}.'
+
+
+@dataclass
+class ContextReorganizationAction(Action):
+    """This action reorganizes the context by providing a structured summary and important files.
+
+    This is useful when:
+    1. The context becomes too large
+    2. The user explicitly requests it
+    3. There's redundant or outdated information (like old file versions)
+
+    Attributes:
+        summary (str): A structured summary of the conversation, containing all important
+            information and insights.
+        files (List[dict]): A list of files from the workspace to add to the context.
+            Each file is represented as a dict with 'path' and optionally 'view_range'.
+        action (str): The action type, namely ActionType.CONTEXT_REORGANIZATION.
+    """
+
+    summary: str = ''
+    files: List[dict] = field(default_factory=list)
+    action: str = ActionType.CONTEXT_REORGANIZATION
+    runnable: ClassVar[bool] = True
+
+    @property
+    def message(self) -> str:
+        file_paths = (
+            ', '.join([f['path'] for f in self.files]) if self.files else 'no files'
+        )
+        return f'Reorganizing context with summary and files: {file_paths}'
