+++ openhands/events/serialization/action.py
@@ -9,6 +9,7 @@ from openhands.events.action.agent import (
     AgentThinkAction,
     ChangeAgentStateAction,
     CondensationAction,
+    ContextReorganizationAction,
     RecallAction,
 )
 from openhands.events.action.browse import BrowseInteractiveAction, BrowseURLAction
@@ -16,7 +17,6 @@ from openhands.events.action.commands import (
     CmdRunAction,
     IPythonRunCellAction,
 )
-from openhands.events.action.context_reorganization import ContextReorganizationAction
 from openhands.events.action.empty import NullAction
 from openhands.events.action.files import (
     FileEditAction,
